+++ lib/bank_scrap/banks/openbank/account.rb
@@ -1,5 +1,7 @@
-module BankScrap
-  class OpenbankAccount < Account
-    attr_accessor :contract_id
+module BankScrap::Banks
+  module Openbank
+    class Account < BankScrap::Account
+      attr_accessor :contract_id
+    end
   end
 end
