+++ lib/bank_scrap/banks/openbank/openbank.rb
@@ -1,0 +1,7 @@
+module BankScrap
+  module Banks
+    module Openbank
+      BankScrap::Banks::Openbank.autoload :Bank,   'bank_scrap/banks/openbank/bank'
+    end
+  end
+end
